refactor: use internal and web folders
