+++ internal/application/routes.go
@@ -3,7 +3,7 @@ package application
 import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
-	"github.com/tombrereton/go-hot-reload/handler"
+	"github.com/tombrereton/go-hot-reload/internal/handler"
 )
 
 func loadRoutes() *chi.Mux {
