+++ web/view/layout/base.templ
@@ -1,6 +1,6 @@
 package layout
 
-import "github.com/tombrereton/go-hot-reload/view/components"
+import "github.com/tombrereton/go-hot-reload/web/view/components"
 
 templ Base(contents templ.Component) {
 	<!DOCTYPE html>
