+++ web/view/marketing/landing.templ
@@ -1,7 +1,7 @@
 package marketing
 
 import (
-	"github.com/tombrereton/go-hot-reload/model"
+	"github.com/tombrereton/go-hot-reload/internal/model"
 )
 
 templ LandingPage(user model.User) {
