+++ web/view/marketing/landing_templ.go
@@ -11,7 +11,7 @@ import "io"
 import "bytes"
 
 import (
-	"github.com/tombrereton/go-hot-reload/model"
+	"github.com/tombrereton/go-hot-reload/internal/model"
 )
 
 func LandingPage(user model.User) templ.Component {
@@ -39,7 +39,7 @@ func LandingPage(user model.User) templ.Component {
 		var templ_7745c5c3_Var3 string
 		templ_7745c5c3_Var3, templ_7745c5c3_Err = templ.JoinStringErrs(user.Name)
 		if templ_7745c5c3_Err != nil {
-			return templ.Error{Err: templ_7745c5c3_Err, FileName: `view/marketing/landing.templ`, Line: 7, Col: 32}
+			return templ.Error{Err: templ_7745c5c3_Err, FileName: `web/view/marketing/landing.templ`, Line: 7, Col: 32}
 		}
 		_, templ_7745c5c3_Err = templ_7745c5c3_Buffer.WriteString(templ.EscapeString(templ_7745c5c3_Var3))
 		if templ_7745c5c3_Err != nil {
